Implemented first-pass for the repository cloning feature.
Currently, the clone feature needed the name of the default branch when pulling the project from the GitHub link. To avoid having to make this a problem for others developers if they have a Git version of >=2.28, I made it something the tool can handle.

For speed, we're focusing right now on downloading the repository from HTTPS. Furthermore, here, we've only taken care of the happy path. We don't display anything to the user yet in case of failure such as the provided link doesn't exist, or they don't have the required access to see the link they provided or even, they're trying to clone the repository to the same location it's been downloaded.

When cloning the repository, the developer will need to provide the name of the folder on which we'll copy the contents from the repository.