Refactored the responsibility of the SCM outside of the Program source file.
The SCM infrastructure would be the main location for commands that we'd like a given source control manager to execute. It could very become quite large if we put everything in there. To end, we could have specific implementation for each SCM we'd like to support and use this infrastructure file as a public interface with which we could invoke internal commands.