Implemented the CLI command allowing to build a player
We have here an infrastructure that can initially check whether the provided build target is among the list of targets that Unity supports before setting it.

Afterwards, we'll try to use the script file in the game project which holds the logic on how to build the player. In this case, I've chosen to go with Android and updated the project locally by adding a script over there thanks to the documentation at

https://docs.unity3d.com/Manual/EditorCommandLineArguments.html?fbclid=IwAR1HYTOhBBkJ838A9bxNtYt4mNqkT0IeRB5OwCCbXUjVnSOnAooM75xKPs4

(Examples)